feat: suspense streaming
